Firefox: Removing SDK extension tests and generalizing selenium tests to support FF
