+++ .gitmodules
@@ -1,6 +1,3 @@
-[submodule "addon-sdk"]
-	path = addon-sdk
-	url = https://github.com/mozilla/addon-sdk.git
 [submodule "translations"]
 	path = translations
 	url = https://git.torproject.org/translation.git
