+++ test/chromium.sh
@@ -39,5 +39,5 @@ else
 	./makecrx.sh
 	echo "running tests"
 	CRX_NAME="`ls -tr pkg/*.crx | tail -1`"
-	$XVFB_RUN python2.7 test/chromium/script.py $CRX_NAME
+	$XVFB_RUN python2.7 test/script.py Chrome $CRX_NAME
 fi
