--- test/firefox/package.json
@@ -1,9 +1,0 @@
-{
-  "name": "https-everywhere-tests",
-  "title": "https-everywhere-tests",
-  "id": "jid1-we5BQOfc6skSMg",
-  "description": "a basic add-on",
-  "author": "",
-  "license": "MPL 2.0",
-  "version": "0.1"
-}
